requirements.txt
streamlit
google-generativeai
phidata
duckduckgo-search